consolidate handler to a single base class, and provide handlers for expected error types.
