+++ src/fastapi_problem/handler/starlette.py
@@ -1,142 +1,19 @@
 from __future__ import annotations
 
-import http
 import logging
-import typing
+import typing as t
 
 from starlette.exceptions import HTTPException
-from starlette.middleware.cors import CORSMiddleware
-from starlette.responses import JSONResponse
 
-from fastapi_problem.error import Problem, StatusProblem
-from fastapi_problem.handler.util import convert_status_code
+from fastapi_problem.handler.base import ExceptionHandler, cors_wrapper_factory, http_exception_handler
 
-if typing.TYPE_CHECKING:
+if t.TYPE_CHECKING:
     from starlette.applications import Starlette
-    from starlette.requests import Request
 
     from fastapi_problem.cors import CorsConfiguration
+    from fastapi_problem.error import StatusProblem
 
 logger_ = logging.getLogger(__name__)
-
-
-def cors_wrapper_factory(
-    cors: CorsConfiguration,
-    handler: typing.Callable[[Request, Exception], JSONResponse],
-) -> typing.Callable[[Request, Exception], JSONResponse]:
-    def wrapper(request: Request, exc: Exception) -> JSONResponse:
-        response = handler(request, exc)
-
-        # Since the CORSMiddleware is not executed when an unhandled server exception
-        # occurs, we need to manually set the CORS headers ourselves if we want the FE
-        # to receive a proper JSON 500, opposed to a CORS error.
-        # Setting CORS headers on server errors is a bit of a philosophical topic of
-        # discussion in many frameworks, and it is currently not handled in FastAPI.
-        # See dotnet core for a recent discussion, where ultimately it was
-        # decided to return CORS headers on server failures:
-        # https://github.com/dotnet/aspnetcore/issues/2378
-        origin = request.headers.get("origin")
-
-        if origin:
-            # Have the middleware do the heavy lifting for us to parse
-            # all the config, then update our response headers
-            mw = CORSMiddleware(
-                app=None,
-                allow_origins=cors.allow_origins,
-                allow_credentials=cors.allow_credentials,
-                allow_methods=cors.allow_methods,
-                allow_headers=cors.allow_headers,
-            )
-
-            # Logic directly from Starlette"s CORSMiddleware:
-            # https://github.com/encode/starlette/blob/master/starlette/middleware/cors.py#L152
-
-            response.headers.update(mw.simple_headers)
-            has_cookie = "cookie" in request.headers
-
-            # If request includes any cookie headers, then we must respond
-            # with the specific origin instead of "*".
-            if mw.allow_all_origins and has_cookie:
-                response.headers["Access-Control-Allow-Origin"] = origin
-
-            # If we only allow specific origins, then we have to mirror back
-            # the Origin header in the response.
-            elif not mw.allow_all_origins and mw.is_allowed_origin(origin=origin):
-                response.headers["Access-Control-Allow-Origin"] = origin
-                response.headers.add_vary_header("Origin")
-
-        return response
-
-    return wrapper
-
-
-def exception_handler_factory(
-    logger: logging.Logger,
-    unhandled_wrappers: dict[str, type[StatusProblem]],
-    *,
-    strip_debug: bool = False,
-    strip_debug_codes: list[int] | None = None,
-) -> typing.Callable[[Exception], JSONResponse]:
-    unhandled_wrappers = unhandled_wrappers or {}
-    strip_debug_codes = strip_debug_codes or []
-
-    def exception_handler(_request: Request, exc: Exception) -> JSONResponse:
-        wrapper = unhandled_wrappers.get("default")
-        ret = (
-            wrapper(str(exc))
-            if wrapper
-            else Problem(
-                title="Unhandled exception occurred.",
-                details=str(exc),
-                code="unhandled-exception",
-            )
-        )
-        headers = {}
-
-        if isinstance(exc, HTTPException):
-            wrapper = unhandled_wrappers.get(str(exc.status_code))
-            title, code = convert_status_code(exc.status_code)
-            details = exc.detail
-            ret = (
-                wrapper(details)
-                if wrapper
-                else Problem(
-                    title=title,
-                    code=code,
-                    details=details,
-                    status=exc.status_code,
-                )
-            )
-            headers = exc.headers or headers
-
-        if isinstance(exc, Problem):
-            ret = exc
-
-        if ret.status >= http.HTTPStatus.INTERNAL_SERVER_ERROR:
-            logger.exception(ret.title, exc_info=(type(exc), exc, exc.__traceback__))
-
-        strip_debug_ = strip_debug or ret.status in strip_debug_codes
-
-        if strip_debug_ and (ret.details or ret.extras):
-            msg = "Stripping debug information from exception."
-            logger.debug(msg)
-
-            for k, v in {
-                "details": ret.details,
-                **ret.extras,
-            }.items():
-                msg = f"Removed {k}: {v}"
-                logger.debug(msg)
-
-        headers["content-type"] = "application/problem+json"
-
-        return JSONResponse(
-            status_code=ret.status,
-            content=ret.marshal(strip_debug=strip_debug_),
-            headers=headers,
-        )
-
-    return exception_handler
 
 
 def generate_handler(
@@ -146,10 +23,13 @@ def generate_handler(
     *,
     strip_debug: bool = False,
     strip_debug_codes: list[int] | None = None,
-) -> typing.Callable:
-    handler = exception_handler_factory(
+) -> t.Callable:
+    handler = ExceptionHandler(
         logger=logger,
         unhandled_wrappers=unhandled_wrappers,
+        handlers={
+            HTTPException: http_exception_handler,
+        },
         strip_debug=strip_debug,
         strip_debug_codes=strip_debug_codes,
     )
